Separate MigrationVrmSpringBone
