+++ Assets/VRM10/Runtime/Migration/MigrationVrmSpringBone.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 885e9b1534dc2ea49ac8e50e519d4b15
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
